Flexible Quorums (#91)
* Initial flexible quorums feature. Fixes TestSystem::get_elected_leader() deadlock

* Fixes flexible BLE. Updates flexible quorum field names and docs.

* Finish renaming fields and add non-optimal quorum overlap warning

* Use flexible_quorum field in configs and create Quorum enums

* Adds FlexibleQuorum struct

* Leader election now favors priority and connectivity

* Leadership connectivity rework

* Add warning to priority API

* Makes BLEBallot private

* Removes both BLEBallot and ballot ordering by connectivity

* Makes Connectivity private

* Updated book

* Doc error

* Updated book

* Removes initial leader config. Removes fail_recovery()

* handle_seq_num returns MessageStatus

* Updated docs

+++ omnipaxos_storage/src/persistent_storage.rs
@@ -34,7 +34,7 @@ const SNAPSHOT: &[u8] = b"SNAPSHOT";
 struct BallotStorage {
     config_id: u32,
     n: u32,
-    priority: u64,
+    priority: u32,
     pid: u64,
 }
